Update Dockerfile
We need to specify the right version of Gradle, otherwise it will break. Using only the "jdk11" to reference the base image is not a best practice, since this can be changed in the future when new Gradle versions are available. I have tested with Gradle version 6.8.3 and was able to get the image built.